requirements.txt
celery[redis]
celery[beat]
flower
praw
prometheus_client
psycopg2-binary
fasttext
pymongo
langdetect
pandas
pyspark
poetry
ariadne
uvicorn
protobuf
grpcio
grpcio-tools 
fastapi